Move Helm references
See #12654

The Kubernetes Access section includes guides for using the Kubernetes
Service as well as guides for running the Auth and Proxy Service on
Kubernetes. This is misleading, since (a) you can run the Auth/Proxy
on Kubernetes without using Kubernetes Access and (b) you can use
Kubernetes Access without running the Auth/Proxy on Kubernetes.

This change focuses on our Helm chart references. These are not
exclusive to the Teleport Kubernetes Service, so I am suggesting moving
these guides to the /docs/setup section to make them more discoverable.