Reorganize the "Home" section of the docs
The organization of the "Home" section of the docs currently says little
about its purpose, making it more difficult for new visitors to the docs
to find content. I've proposed changing this to a "Get Started" section
that includes content related to trying out Teleport and choosing an
edition.

(Re)move pages that are unrelated to getting started:

- Rename the "Preview" section to "Releases" and move the Changelog
  there. The Preview section doesn't include preview content very often,
  and usually only includes the Upcoming Releases page. I think it makes
  sense to devote this section to releases.

- Remove the Adopters page and redirect to the Teleport case studies
  page. The Adopters page receives close to no traffic and is
  infrequently updated, so keeping it risks giving users outdated
  information.

- Remove the Production guide. Redirects currently make it impossible
  to navigate to this page.

- Move the Digital Ocean guide to "/docs/setup/deployments", since other
  platform-specific deployment guides are also placed there.